rgw: svc: initial work on zone_svc
